pyproject.toml
[project]
name = "imap-idle-listener"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aioimaplib>=2.0.1",
    "html2text>=2024.2.26",
    "httpx>=0.28.1",
    "python-dotenv>=1.1.0",
]
